refactor: npm i sass, rename all .css to .scss
