+++ tripal-traid-game/package.json
@@ -10,6 +10,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
+    "sass": "^1.50.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
