+++ tripal-traid-game/src/components/Container/Container.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import cn from "classnames";
-import s from "./Container.module.css";
+import s from "./Container.module.scss";
 
 const Container = ({ children, className }) => {
   return <div className={cn(s.root, className)}>{children}</div>;
