+++ tripal-traid-game/src/components/Footer/Footer.module.scss
