+++ tripal-traid-game/src/components/Footer/index.js
@@ -1,18 +1,3 @@
-import Container from '../Container';
-import style from './Footer.module.css';
-
-const Footer = () => {
-  return (
-    <footer className={style.root}>
-      <Container className={style.container}>
-        <div className={style.footerWrap}>
-          Coded with
-          <span className={style.heart}></span>
-          by You
-        </div>
-      </Container>
-    </footer>
-  )
-};
+import Footer from "./Footer";
 
 export default Footer;
