+++ tripal-traid-game/src/components/Heading/Heading.module.scss
