+++ tripal-traid-game/src/components/Slider/Slider.module.scss
