+++ tripal-traid-game/src/components/Text/Text.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import cn from "classnames";
-import s from "./Text.module.css";
+import s from "./Text.module.scss";
 
 const Text = (props) => {
   const {
