+++ tripal-traid-game/src/components/Text/Text.module.scss
